requirements.txt
fastapi==0.95.0
uvicorn==0.20.0
pillow==9.5.0
python-dotenv==1.0.0